+++ pds_pipelines/PDSinfo.json
@@ -124,7 +124,7 @@
     },
     "mro_hirise_rdr": {
         "archiveid": "125",
-        "path": "/pds_san/PDS_Archive/Mars_Reconnaissance_Orbiter/HiRISE/RDR",
+        "path": "/pds_san/PDS_Archive/Mars_Reconnaissance_Orbiter/HiRISE/",
 	"bandbinQuery": "Name",
 	"upc_reqs": [".LBL"]
     },
