+++ recipe/new/mro_hirise_rdr.json
@@ -5,7 +5,15 @@
         "recipe": {
             "isis.hirdr2isis": {
                 "from_": "{{inputfile}}",
-                "to": "{{cam_info_file}}"
+                "to": "{{no_extension_inputfile}}.cub"
+            },
+            "isis.caminfo": {
+                "from_": "{{no_extension_inputfile}}.cub",
+                "to": "{{cam_info_file}}",
+                "format": "pvl",
+                "isislabel": "true",
+                "originallabel": "true",
+                "geometry": "false"
             },
             "gdal_translate": {
                 "scaleParams": [[1,65535,1,1]],
@@ -18,12 +26,12 @@
             },
             "gdal_polygonize": {
                 "input_file": "{{no_extension_inputfile}}.tif",
-                "output_name": "{{footprint_file}}"
+                "output_name": "{{no_extension_inputfile}}_footprint.shp"
             },
             "ogr2ogr": {
                  "format": "GeoJSON",
                  "dstSRS": "+proj=longlat +R=3396190 +no_defs",
-                 "dest": "{{no_extension_inputfile}}_footprint.json",
+                 "dest": "{{footprint_file}}",
                  "src": "{{no_extension_inputfile}}_footprint.shp"
             }
         },
